Cambios - Franco
Backend del apartado del administrador

+++ src/routes/admin.js
@@ -14,7 +14,7 @@ router.get('/usuarios', (req,res,next) => {
     if(req.isAuthenticated()){
         return next();
     }
-        res.redirect('/login');
+        res.redirect('/');
     },(req,res) =>{
     conn.query('SELECT * FROM usuario', (err, usuarios) => {
         if (err) {
@@ -26,12 +26,18 @@ router.get('/usuarios', (req,res,next) => {
     });
 });
 
-router.get('/eliminar/:correo', (req, res) => {
-    const { correo } = req.params;
-    req.getConnection((err, conn) => {
+router.get('/eliminar/:correo', (req,res,next) => {
+    if(req.isAuthenticated()){
+        return next();
+    }
+        res.redirect('/');
+    },(req,res) =>{
+        const { correo} = req.params;
         conn.query('DELETE FROM usuario WHERE correo = ?', [correo], (err, rows) => {
-            res.redirect('/admin');
-        });
+        if (err) {
+            res.json(err);
+        }
+        res.redirect('/usuarios');
     });
 });
 
@@ -44,25 +50,79 @@ router.get('/eliminar/:correo', (req, res) => {
 
     // SERVICIOS REGIONALES
 
-    router.get('/servicios-regionales', (req, res) => {
-        req.getConnection((err, conn) =>{
-            conn.query('SELECT * FROM servicio_admin', (err, servicios) => {
-                if (err) {
-                    res.json(err);
-                }
-                res.render('servicios-reg', {
-                    data: servicios
-                });
+    router.get('/servicios-regionales', (req,res,next) => {
+        if(req.isAuthenticated()){
+            return next();
+        }
+            res.redirect('/');
+        },(req,res) =>{
+        conn.query('SELECT * FROM servicio_admin', (err, servicios) => {
+            if (err) {
+                res.json(err);
+            }
+            res.render('servicios-reg', {
+                data: servicios
             });
         });
     });
 
-    router.post('/agregar-servicio', (req, res) => {
+    router.post('/agregar-servicio', (req, res, next) => {
+        if(req.isAuthenticated()){
+            return next();
+        }
+            res.redirect('/');
+        },(req,res) =>{
         const data = Object.assign({},req.body)
-        req.getConnection((err, conn) => {
-            conn.query('INSERT INTO servicio_admin set ?', [data], (err, customer) => {
-                res.redirect('/servicios-regionales');
+        conn.query('INSERT INTO servicio_admin set ?', [data], (err, customer) => {
+            res.redirect('/servicios-regionales');
+        });
+    });
+
+    router.get('/eliminar-servicio-reg/:id_servicio', (req, res, next) => {
+        if(req.isAuthenticated()){
+            return next();
+        }
+            res.redirect('/');
+        },(req,res) =>{
+            const { id_servicio } = req.params;
+            conn.query('DELETE FROM servicio_admin WHERE id_servicio = ?', [id_servicio], (err, rows) => {
+            if (err) {
+                res.json(err);
+            }
+            res.redirect('/servicios-regionales');
+        });
+    });
+    
+    router.get('/modificar-servicio-reg/:id_servicio', (req, res, next) => {
+        if(req.isAuthenticated()){
+            return next();
+        }
+            res.redirect('/');
+        },(req,res) =>{
+            const { id_servicio } = req.params;
+            conn.query('SELECT * FROM servicio_admin WHERE id_servicio = ?', [id_servicio], (err, rows) => {
+            if (err) {
+                res.json(err);
+            }
+            res.render('mod-servicios-reg', {
+                data: rows[0]
             });
+        });
+    });
+    
+    router.post('/modificar-servicio-reg/:id_servicio', (req, res, next) => {
+        if(req.isAuthenticated()){
+            return next();
+        }
+            res.redirect('/');
+        },(req,res) =>{
+            const { id_servicio } = req.params;
+            const newServicio = Object.assign({},req.body)
+            conn.query('UPDATE servicio_admin set ? WHERE id_servicio = ?', [newServicio, id_servicio], (err, rows) => {
+            if (err) {
+                res.json(err);
+            }
+            res.redirect('/servicios-regionales');
         });
     });
 
@@ -72,15 +132,19 @@ router.get('/eliminar/:correo', (req, res) => {
 
     // SERVICIOS DE USUARIOS
 
-    router.get('/servicios-usuarios', (req, res) => {
-        req.getConnection((err, conn) =>{
-            conn.query('SELECT * FROM servicio_emprendedor', (err, servicios) => {
-                if (err) {
-                    res.json(err);
-                }
-                res.render('servicios-user', {
-                    data: servicios
-                });
+
+    router.get('/servicios-usuarios', (req,res,next) => {
+        if(req.isAuthenticated()){
+            return next();
+        }
+            res.redirect('/');
+        },(req,res) =>{
+        conn.query('SELECT * FROM servicio_admin', (err, servicios) => {
+            if (err) {
+                res.json(err);
+            }
+            res.render('servicios-user', {
+                data: servicios
             });
         });
     });
@@ -91,27 +155,89 @@ router.get('/eliminar/:correo', (req, res) => {
 
 /* -------------------------------------- */
 
-//
-router.post('/agregar', (req, res) => {
-    const data = Object.assign({},req.body)
-    console.log(req.body);
-    req.getConnection((err, conn) => {
-        conn.query('INSERT INTO usuario set ?', [data], (err, customer) => {
-            res.redirect('/admin');
+// Apartado de ATRACTIVOS REGIONALES
+
+router.get('/atractivos-regionales', (req,res,next) => {
+    if(req.isAuthenticated()){
+        return next();
+    }
+        res.redirect('/');
+    },(req,res) =>{
+    conn.query('SELECT * FROM atractivo_admin', (err, atractivos) => {
+        if (err) {
+            res.json(err);
+        }
+        res.render('atractivos-reg', {
+            data: atractivos
         });
     });
 });
 
+router.post('/agregar-atractivo', (req,res,next) => {
+    if(req.isAuthenticated()){
+        return next();
+    }
+        res.redirect('/');
+    },(req,res) =>{
+    conn.query('INSERT INTO atractivo_admin set ?', [req.body], (err, resp) => {
+        if (err) {
+            res.json(err);
+        }
+        res.redirect('/atractivos-regionales');
+    });
+});
 
-router.get('/servicios', (req, res) => {
-    const { id } = req.params;
-    req.getConnection((err, conn) => {
-        conn.query('SELECT * FROM customer', (err, customer) => {
-            res.render('admin-servicios', {
-                data: customer
-            });
+router.get('/eliminar-atractivo-reg/:id_atractivo', (req, res, next) => {
+    if(req.isAuthenticated()){
+        return next();
+    }
+        res.redirect('/');
+    },(req,res) =>{
+        const { id_atractivo } = req.params;
+        conn.query('DELETE FROM atractivo_admin WHERE id_atractivo = ?', [id_atractivo], (err, rows) => {
+        if (err) {
+            res.json(err);
+        }
+        res.redirect('/atractivos-regionales');
+    });
+});
+
+router.get('/modificar-atractivo-reg/:id_atractivo', (req, res, next) => {
+    if(req.isAuthenticated()){
+        return next();
+    }
+        res.redirect('/');
+    },(req,res) =>{
+        const { id_atractivo } = req.params;
+        conn.query('SELECT * FROM atractivo_admin WHERE id_atractivo = ?', [id_atractivo], (err, rows) => {
+        if (err) {
+            res.json(err);
+        }
+        res.render('mod-atractivos-reg', {
+            data: rows[0]
         });
     });
 });
+
+router.post('/modificar-atractivo-reg/:id_atractivo', (req, res, next) => {
+    if(req.isAuthenticated()){
+        return next();
+    }
+        res.redirect('/');
+    },(req,res) =>{
+        const { id_atractivo } = req.params;
+        const newAtractivo = Object.assign({},req.body)
+        conn.query('UPDATE atractivo_admin set ? WHERE id_atractivo = ?', [newAtractivo, id_atractivo], (err, rows) => {
+        if (err) {
+            res.json(err);
+        }
+        res.redirect('/atractivos-regionales');
+    });
+});
+
+// Fin apartado de ATRACTIVOS REGIONALES
+
+/* -------------------------------------- */
+
 
 module.exports = router;
